+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [App4.0.1] - 2025-11-18
+## [App4.0.1] [SequenceCreator1.3.1]- 2025-11-18
 ### Added
 - Device connection status tracking: Real-time monitoring of 11 device types based on event history.
 - Autofocus event tracking: Dedicated store with real-time autofocus state, points, and HFR visualization.
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fixed Camera settings layout
+- SequenceCreator Autofocus sequence fixed
+- Fixed Dialog Modal
 
 ## [Plugin1.2.1.0] [App4.0.0] [webcam-1.0.1] [LivestackPlugin0.4.1] [logfile-collector1.0.2] - 2025-11-16
 ### Important information